kineteco-research/requirements.txt
alembic==1.0.5
atomicwrites==1.2.1
attrs==18.2.0
Click==7.0
coverage==4.5.2
flake8==3.6.0
Flask==1.0.2
Flask-Migrate==2.3.1
flask-restplus==0.12.1
Flask-Script==2.0.6
flask-shell-ipython==0.4.0
Flask-SQLAlchemy==2.3.2
ipython==7.2.0
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==4.3.0
mypy==0.641
mypy-extensions==0.4.1
pact-python==0.18.0
pluggy==0.8.0
psycopg2==2.7.6.1
py==1.7.0
pycodestyle==2.4.0
pyflakes==2.0.0
pytest==4.0.1
pytest-cov==2.6.0
python-dateutil==2.7.5
python-editor==1.0.3
requests==2.21.0
six==1.11.0
SQLAlchemy==1.2.14
typed-ast==1.1.0
Werkzeug==0.14.1
